fixes #9156
node: false should not disable CommonJs features
