+++ lib/WebpackOptionsApply.js
@@ -21,6 +21,7 @@ const RecordIdsPlugin = require("./RecordIdsPlugin");
 
 const APIPlugin = require("./APIPlugin");
 const ConstPlugin = require("./ConstPlugin");
+const CommonJsStuffPlugin = require("./CommonJsStuffPlugin");
 const CompatibilityPlugin = require("./CompatibilityPlugin");
 
 const TemplatedPathPlugin = require("./TemplatedPathPlugin");
@@ -293,6 +294,7 @@ class WebpackOptionsApply extends OptionsApply {
 			const NodeStuffPlugin = require("./NodeStuffPlugin");
 			new NodeStuffPlugin(options.node).apply(compiler);
 		}
+		new CommonJsStuffPlugin().apply(compiler);
 		new APIPlugin().apply(compiler);
 		new ConstPlugin().apply(compiler);
 		new UseStrictPlugin().apply(compiler);
