+++ test/configCases/parsing/issue-9156/webpack.config.js
@@ -1,0 +1,4 @@
+module.exports = {
+	target: "web",
+	node: false
+};
